Add reports functionality with routes, PDF generation, and seeding for policies

+++ database/migrations/2025_03_14_065556_create_policies_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('policies', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('code')->unique()->nullable();
+            $table->bigInteger('code')->unique()->nullable();
             $table->timestamps();
         });
     }
